[Refactor] 1. Rename CommonExt to Common 2. Move property wrapper @Clamping to Common/PropertyWrapper
