+++ Sources/Rings/KnobComponents/Layers/ArcKnobLayer.swift
@@ -6,6 +6,7 @@
 //
 
 import SwiftUI
+import Common
 
 public struct ArcKnobLayer : KnobLayer {
     public var isFixed: Bool = false
