+++ Sources/Rings/RingText.swift
@@ -1,6 +1,6 @@
 import SwiftUI
 import CoreGraphicsExtension
-import CommonExts
+import Common
 
 @available(iOS 13.0, macOS 10.15, tvOS 13.0, watchOS 6.0, *)
 public extension CGAngle {
